refactor(backend): remove pingInterval loop setting latencies for players
We shouldn't send to a go channel in a goroutine that isn't the reciever
of that channel since it can cause an instance where that channel is
closed and we have a gorountine still wanting to send to the channel.

I.E

client.writePump (owns channel should be able to close it)

playerClient.pingInterval (outside and has chances to cause panic if
sends to closed channel)

I am opting to NOT close the client.send channel manually since I believe when
the client eventually disconnects and  the room Hub also disconnects the
go-lang garbage collector will close the channel when needed.